requirements.txt
torch >= 2.0.0
transformers>=4.36.0
datasetsr>=2.15.0
accelerate>=0.25.0
deepspeed>=0.12.6
wandb>=0.16.1
matplotlib>=3.8.2
seaborn>=0.13.0
pandas>=2.1.4
numpy>=1.24.4
scikit-learn>=1.3.2
scipy>=1.11.4
plotly>=5.17.0
faker>=20.1.0
jinja2>=3.1.2
pyyaml>=6.0.1
tqdm>=4.66.1
